// src/app/Models/caso.ts
export interface Caso {
  accountNumber: string;
  action: string;
  addressFails: string;
  addressedTo: string;
  adviser: string;
  applicantName: string;
  applicantSurname: string;
  baseStations: string;
  caseAssignmentGroup: string;
  caseDateEnd: string;
  caseHour: string;
  caseHourEnd: string;
  cause: string;
  ccmm: string;
  channel: string;
  city: string;
  comments: string;
  contactNumber: string;
  customerIdentification: string;
  dateCase: string;
  dateMod: string;
  department: string;
  diagnosisLevel: string;
  diagnosticGuide: string;
  email: string;
  equipmentModel: string;
  factsRequest: string;
  failureDate: string;
  failureDateSolution: string;
  functionality: string;
  id: number;
  idCase: string;
  idCategory: string;
  idCloseCode: string;
  idCustomerVoice: string;
  idPublication: string;
  idState: number;
  idSubcategory: string;
  impact: string;
  incidentAssignmentGroup: string;
  incidentManagementSystem: string;
  latitude: string;
  lineType: string;
  longitude: string;
  msisdnFailure: string;
  namePersonReport: string;
  node: string;
  notes: string;
  originCase: string;
  personReport: string;
  phoneNumber: string;
  planType: string;
  priority: string;
  problem: string;
  product: string;
  reason: string;
  relatedChange: string;
  relatedIncident: string;
  remedy: string;
  request: string;
  requestObject: string;
  sector: string;
  service: string;
  sla: string;
  status: string;
  statusReason: string;
  subservice: string;
  technology: string;
  timeByState: string;
  transaction: string;
  typeRequest: string;
  typificationCloseCase: string;
  urgency: string;
  userMod: string;
  zone: string;
  zoneType: string;
}